requirements.txt
black==24.10.0
click==8.1.7
Flask==2.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
Werkzeug==2.0.3
